generalizing nodejs into node+browser
