+++ __tests__/js/test_project/src/main.js
@@ -1,7 +1,9 @@
 //import * as IPFS from 'ipfs-core';
-import ipfsClient from 'ipfs-http-client';
+//import ipfsClient from 'ipfs-http-client';
 //import * as IPFS from 'ipfs-core/src/index.js';
 import { helloWorld } from './hello';
 
 //import Core from "../../../../decelium_wallet/core.js";
+
+
 document.getElementById('hello-message').innerText = helloWorld();
